Renamed package runtime to diag
Since it deals with diagnostics and we want to use the name runtime for
a package in azcore.
Deleted FrameError as it's not very useful.
Added Caller() for times when a full stack trace isn't needed.
